Enable runtime renderer switching between DirectX 11 and OpenGL
Replace compile-time USE_OPENGL define with HAS_OPENGL/HAS_DX11 so both
backends are compiled into a single binary. The renderer is now selected
at startup via configuration file (RendererAPI setting), command line
(-api dx11/opengl), or the new in-game Display Settings menu option.

Main changes:
- Remove Debug_OpenGL/Release_OpenGL build configurations from sln/vcxproj
- Add GraphicsAPI enum (Auto/DirectX11/OpenGL) and runtime factory in
  Renderer::Create()
- Extract Windows-specific GL code into GLPlatformHelpers using SDL3
  platform macros
- Replace USE_OPENGL with HAS_OPENGL in 27 GL files, SDL_PLATFORM_WIN32
  with HAS_DX11 in 26 DX11 files
- Move HLSL shaders to Shaders/HLSL/ subdirectory (GLSL already in
  Shaders/GLSL/)
- Add NeedsFBOYFlip() virtual for runtime backend detection instead of
  preprocessor checks
- Add renderer option to Display Settings menu with Auto/DX11/OpenGL
  cycling
- Reorder engine initialization so config loads before renderer creation

Bug fixes:
- Fix OpenGL crash (access violation at 0x0) caused by GetAdapterInfo()
  calling glGetString() before GLAD was loaded; defer adapter info query
  until after GL context creation in Initialize()
- Fix InitDefaultConfiguration() crash by removing premature call to
  g_Renderer.GetDefaultAdapterName() before renderer exists
- Fix animated textures: integer arithmetic in GetFrame() caused wrong
  frame calculation, corrected to float math
- Fix HUD bar shader: was using flat vertex color instead of sampling
  the bar texture with proper UVs
- Fix sky shader: replace incorrect Item CB references (Color/World)
  with Sky CB uniforms (SkyColor/SkyWorld), remove unused CBItem include

+++ TombEngine/Shaders/GLSL/AnimatedTextures.glsl
@@ -57,7 +57,7 @@ vec2 CalculateUVRotateForLegacyWaterfalls(vec2 uv, uint frame)
 vec2 GetFrame(uint index, uint offset)
 {
     float speed = float(FPS) / 30.0;
-    int frame = int(Frame * uint(speed) + offset) % int(NumAnimFrames);
+    int frame = int(float(Frame) * speed + float(offset)) % int(NumAnimFrames);
 
     vec2 result = vec2(0);
